Docs/env: usar SESSION_DRIVER=cookie y habilitar SESSION_PARTITIONED_COOKIE=true para cookies de terceros (CHIPS)

+++ Back-E/config/mail.php
@@ -41,7 +41,8 @@ return [
             'transport' => 'smtp',
             'url' => env('MAIL_URL'),
             'host' => env('MAIL_HOST', '127.0.0.1'),
-            'port' => env('MAIL_PORT', 2525),
+            // Cast explícito a int para evitar errores de tipo en Symfony\Component\Mailer\Transport\Dsn
+            'port' => (int) env('MAIL_PORT', 2525),
             'encryption' => env('MAIL_ENCRYPTION', 'tls'),
             'username' => env('MAIL_USERNAME'),
             'password' => env('MAIL_PASSWORD'),
